+++ Source/BackspaceDetectingTextField.swift
@@ -1,0 +1,36 @@
+//
+//  BackspaceDetectingTextField.swift
+//  WSTagsField
+//
+//  Created by Ilya Seliverstov on 11/07/2017.
+//  Copyright © 2017 Whitesmith. All rights reserved.
+//
+
+import UIKit
+
+protocol BackspaceDetectingTextFieldDelegate: UITextFieldDelegate {
+    /// Notify whenever the backspace key is pressed
+    func textFieldDidDeleteBackwards(_ textField: UITextField)
+}
+
+class BackspaceDetectingTextField: UITextField {
+    
+    var onDeleteBackwards: Optional<()->()>
+    
+    init() {
+        super.init(frame: CGRect.zero)
+    }
+    
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    override func deleteBackward() {
+        if let deleteBackwardsEvent = onDeleteBackwards {
+            deleteBackwardsEvent()
+        }
+        // Call super afterwards. The `text` property will return text prior to the delete.
+        super.deleteBackward()
+    }
+    
+}
